feat(MatchRoute): Moved traversal middleware opportunity to after a route is matched
Traversal middleware now receives the matched route, a preview of route params, and whether or not traversal plans on terminating on that route. Traversal middleware can null out the route to prevent traversal to stop traversing that route. It can mark the candidate as terminal to force traversal to select that route. Lastly, if the route is marked as terminal it can modify the final route params