fix(MatchRoute): Give favor to the order in which routes are defined
Previous algorithm would have given favor to the route that finished first. This would be problematic if you had a long running guard on a root route. The router would have selected 404 instead of that route because your 404 route would finish first. Switching to concat instead of merge fixes this problem.