feat(RouteTraverser): Allow forward slash in nested route paths
Before we ignored nested child routes that contained a forward slash at the beginning of the path. This was to potentially allow for absolute path URLs in nested routes, however there is no way to make this feature work with async route configuration. This change removes this restriction and allows you to add a forward slash at the beginning of the path with no functional difference.